moved Views trait into CrudPanel instead of CrudController
