+++ src/app/Http/Controllers/CrudFeatures/Reorder.php
@@ -25,7 +25,7 @@ trait Reorder
         $this->data['title'] = trans('backpack::crud.reorder').' '.$this->crud->entity_name;
 
         // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view('crud::reorder', $this->data);
+        return view($this->crud->getReorderView(), $this->data);
     }
 
     /**
