+++ src/app/Http/Controllers/CrudFeatures/ShowDetailsRow.php
@@ -21,6 +21,6 @@ trait ShowDetailsRow
         $this->data['crud'] = $this->crud;
 
         // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->getDetailsRowView(), $this->data);
+        return view($this->crud->getDetailsRowView(), $this->data);
     }
 }
